https://github.com/WWBN/AVideo/issues/6954#issuecomment-1234921484

+++ plugin/ReportVideo/Objects/videos_reported.php
@@ -150,7 +150,7 @@ class VideosReported extends ObjectYPT {
     
     public function delete() {
         $deleted = parent::delete();
-        if($deleted){
+        if($deleted && class_exists('Cache')){
             Cache::deleteFirstPageCache();
         }
         return $deleted;
